frappe-oauth/backend/requirements.txt
Flask==2.3.3
requests==2.32.1
python-dotenv==1.0.0
flask-cors==3.0.10
